Componentize packageSelector
